feat: testing powershelled commands
