+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@2501-ai/cli",
-  "version": "0.2.38-alpha.4",
+  "version": "0.2.38-alpha.5",
   "description": "",
   "main": "dist/index.js",
   "bin": {
